init lss connector in broker, cleanup main func
